[cad-enry] for pr
