+++ info_processing/git/helpers.py
@@ -153,3 +153,17 @@ def save_data(data: Dict, path: Path):
     """
     with open(path.resolve(), "w") as f:
         f.write(json.dumps(data, indent=6, ensure_ascii=False))
+
+
+def clone_repo(path: str) -> Repo:
+    """
+    Function cloned repo, if it is not existed, otherwise instantiate the existing one.
+    :param: The path in local directory or url to GitHub repository.
+    :return: Repository instance.
+    """
+    repo_name = path[path.rfind('/') + 1:]
+    path_to_repo = str(Path(f"{Path().cwd().parent}/treesitter/{repo_name}"))
+    parent_path = str(Path(path_to_repo).parent)
+    if not os.path.exists(parent_path):
+        os.makedirs(parent_path)
+    return Repo(path_to_repo) if os.path.exists(path_to_repo) else clone(path, path_to_repo)
